requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
#backports.zoneinfo==0.2.1
#certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.6.1
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==2.0.2
hugchat==0.2.6
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.302
langsmith==0.0.41
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
openai==0.28.1
packaging==23.1
pydantic==2.4.1
psycopg2-binary
pydantic-core==2.10.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.0.5
yarl==1.9.2
